test: move per-project sequence serialization test into configureVitest
Fold the serialized-config assertion into configureVitest.test.ts (reusing its
createVitest helper) instead of a standalone file, and drop the shared-seed
test.
